Fix issue #83: Abort with error message if svnadmin cannot be executed.
Previously, the inability to run svnadmin resulted in a crash in
pass8.  This commit adds a check to main() that svnadmin can be
executed (if the results of the conversion should go into an
archive).  If the check fails, the run is aborted with a (hopefully)
understandable error message.

Caveats:

1. The test for svnadmin is done even if pass8 is not being executed
   in this run.  This is consistent with other, similar checks (for
   the co/cvs executables).

2. To test for the presence of svnadmin, it is run as "svnadmin help".
   This doesn't have any side effects, so it should be OK.  It could
   possibly be a problem if the behavior of svnadmin was different in
   other versions (for example, either not having a "help" subcommand
   or outputting the help to stderr instead of stdout).

* cvs2svn (main): Abort with error message if svnadmin cannot be
  executed.