Store SvnRevisionRange instances for SymbolicNameFillingGuide leaf nodes.
The leaf nodes of SymbolicNameFillingGuide._node_tree used to be
dictionaries with special keys (OPENING_KEY and CLOSING_KEY).  The
only distinction between an intermediate node and a leaf node was the
presence of one of these keys.

This commit changes SymbolicNameFillingGuide to store SvnRevisionRange
objects for leaf nodes.  Dictionaries are still used for intermediate
nodes.

* cvs2svn (SymbolicNameFillingGuide._node_tree): Changed contents as
  described above.

  (SymbolicNameFillingGuide.OPENING_KEY,
  SymbolicNameFillingGuide.CLOSING_KEY): Removed (not needed anymore).

  (SymbolicNameFillingGuide._list_revnums): Return a list of
  SvnRevisionRange objects (including openings and closings) rather
  than the revision numbers themselves.

  (SymbolicNameFillingGuide.get_best_revnum): Pull the openings and
  closings out of the SvnRevisionRange objects returned by
  _list_revnums().

  (SymbolicNameFillingGuide._get_sub_sources,
  SVNRepositoryMirror._fill): Changed to deal with new tree contents.