Don't sort arguments before passing them to cvs2svn.
Previously, arguments were sorted before passing them to cvs2svn.
This was apparently to avoid superfluous conversions in the case of
calling ensure_conversion() twice with the same arguments but in
different order.

However, order-dependent options might be interesting (e.g., see
discussion of proposed --include option), in which case sorting the
arguments would be bad.  And who's going to call ensure_conversion()
twice with the same arguments but in different order?  (Even if this
happens, it only results in a redundant conversion.)

* run-tests.py (ensure_conversion): Pass arguments to cvs2svn in the
  order that they appear in ARGS.