Fix a problem handling error_re arguments.
The error_re argument was being ignored if the command didn't output
anything to stderr.  This changes it to the documented behavior, that
if error_re is set, there *must* be an error.  This commit also fixes
two test cases that were written incorrectly but didn't fail because
of the problem.

* run-tests.py (run_program): Fix handling of error_re argument as
  described above.

  (overlapping_branch, repeated_deltatext): Fix test cases to check
  for the expected warnings on stdout rather than stderr.