Teach destroy_repository.py to destroy only basenames of the given RCS files
This patch adds the --basenames option for destroying filenames, except for
their last component. (The ',v' extension is not regarded as a "proper"
extension, for the purposes of --basenames) For example:
	foo.txt    -> file001.txt
	bar.h      -> file002.h
	bar.h,v    -> file002.h,v
	foo.cpp    -> file001.cpp
	Makefile   -> Makefile
	Makefile,v -> Makefile,v
	a.b.c.d    -> file003.d

The --basenames option is overridden by the --filenames option. For example:
In order to destroy everything except the filename extensions, the following
command-line must be used:
	contrib/destroy_repository.py --all --no-filenames --basenames PATH

Patch by: Johan Herland <johan@herland.net>