Move the main example cvs2git options file to the root directory.
The main cvs2git options file was previously stored at
test-data/main-cvsrepos/cvs2svn-git.options, which is a little bit
obscure.  Move that file to the root directory and rename it to
cvs2git-example.options.

Where the file was originally located, add a stub file that
execfile()s the moved file and rename the stub file to
cvs2git.options.

* cvs2git-example.options: New file, renamed and slightly modified
  from test-data/main-cvsrepos/cvs2svn-git.options.

* test-data/main-cvsrepos/cvs2svn-git.options: Delete file.

* test-data/main-cvsrepos/cvs2git.options: New file, replacing
  test-data/main-cvsrepos/cvs2svn-git.options.  It just execfile()s
  cvs2git-example.options.

* run-tests.py (main_git): Adjust for the renaming of the
  cvs2svn-git.options file.

* www/cvs2git.html: Refer to the cvs2git-examples.options file as the
  main example options file for cvs2git.