Add failing test involving a branch from an excluded vendor branch.
Patch by: Jon Foster <jon.foster@cabot.co.uk>

In the following scenario:

* my-branch sprouts from NTDBR on vendor-branch

* vendor-branch and vendor-tag are being excluded (and therefore the
  parent of my-branch is being grafted to trunk)

it should be possible to use a symbol-hints file to tell cvs2svn to
use trunk as the preferred parent for my-branch.

This test currently fails; cvs2svn refuses to accept trunk as the
preferred parent of my-branch because it is not recognized as a
possible parent.