Handle --default-eol=binary as per documentation.
Setting --default-eol=binary has been broken since r4687.  The error
has presumably gone undetected because the default value was treated
as "binary" (so --default-eof=binary never needed to be used
explicitly) but handled internally as None (which wasn't broken).

The breakage resulted in a crash, so there is no risk of hidden data
corruption.